Documentation and fixes
