+++ lambda/custom/handlers/Channels/AddModeratorIntentHandlers.js
@@ -15,12 +15,14 @@ const StartedAddModeratorIntentHandler = {
 
 		const { attributesManager } = handlerInput;
 		const sessionAttributes = attributesManager.getSessionAttributes() || {};
-
+		// clear the session attributes that this intent will make use of
 		delete sessionAttributes.user;
 		delete sessionAttributes.channel;
 		delete sessionAttributes.similarusers;
 		delete sessionAttributes.similarChannels;
 
+		// this will take care of the cases when both the slots will be filled during intent invocation
+		// eg: make username as moderator of channel channelname
 		if (intentSlots.username.confirmationStatus === 'NONE' && intentSlots.username.value) {
 			return resolveUser(handlerInput, 'username', 'choice');
 		} else if (intentSlots.channelname.confirmationStatus === 'NONE' && intentSlots.channelname.value) {
