+++ lambda/custom/handlers/Channels/AddOwnerIntentHandler.js
@@ -16,11 +16,14 @@ const StartedAddOwnerIntentHandler = {
 		const { attributesManager } = handlerInput;
 		const sessionAttributes = attributesManager.getSessionAttributes() || {};
 
+		// clear the session attributes that this intent will make use of
 		delete sessionAttributes.user;
 		delete sessionAttributes.channel;
 		delete sessionAttributes.similarusers;
 		delete sessionAttributes.similarChannels;
 
+		// this will take care of the cases when both the slots will be filled during intent invocation
+		// eg: make username as owner of channel channelname
 		if (intentSlots.username.confirmationStatus === 'NONE' && intentSlots.username.value) {
 			return resolveUser(handlerInput, 'username', 'choice');
 		} else if (intentSlots.channelname.confirmationStatus === 'NONE' && intentSlots.channelname.value) {
