+++ lambda/custom/handlers/Channels/ArchiveChannelIntentHandler.js
@@ -27,7 +27,14 @@ const ArchiveChannelIntentHandler = {
 				.withSimpleCard(ri('ARCHIVE_CHANNEL.CARD_TITLE'), speechText)
 				.getResponse();
 		} catch (error) {
-			console.error(error);
+			const speechText = ri('SOMETHING_WENT_WRONG');
+			const repromptText = ri('GENERIC_REPROMPT');
+
+			return handlerInput.jrb
+				.speak(speechText)
+				.speak(repromptText)
+				.reprompt(repromptText)
+				.getResponse();
 		}
 	},
 };
