+++ lambda/custom/handlers/Channels/GetLastMessageFromChannelIntentHandler.js
@@ -312,7 +312,14 @@ const GetLastMessageFromChannelIntentHandler = {
 			}
 
 		} catch (error) {
-			console.error(error);
+			const speechText = ri('SOMETHING_WENT_WRONG');
+			const repromptText = ri('GENERIC_REPROMPT');
+
+			return handlerInput.jrb
+				.speak(speechText)
+				.speak(repromptText)
+				.reprompt(repromptText)
+				.getResponse();
 		}
 	},
 };
