+++ lambda/custom/handlers/Channels/GetUnreadsIntentHandler.js
@@ -23,7 +23,14 @@ const GetUnreadsIntentHandler = {
 				.reprompt(repromptText)
 				.getResponse();
 		} catch (error) {
-			console.error(error);
+			const speechText = ri('SOMETHING_WENT_WRONG');
+			const repromptText = ri('GENERIC_REPROMPT');
+
+			return handlerInput.jrb
+				.speak(speechText)
+				.speak(repromptText)
+				.reprompt(repromptText)
+				.getResponse();
 		}
 	},
 };
