--- lambda/custom/handlers/Channels/PrivateChannelIntents.js
@@ -1,75 +1,0 @@
-const { ri } = require('@jargon/alexa-skill-sdk');
-const helperFunctions = require('../../helperFunctions');
-
-const PostGroupEmojiMessageIntentHandler = {
-	canHandle(handlerInput) {
-		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
-			handlerInput.requestEnvelope.request.intent.name === 'PostGroupEmojiMessageIntent';
-	},
-	async handle(handlerInput) {
-		try {
-			const {
-				accessToken,
-			} = handlerInput.requestEnvelope.context.System.user;
-
-			const messageData = handlerInput.requestEnvelope.request.intent.slots.groupemojimessage.value;
-			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.groupmessageemojichannelname.value;
-			const channelName = helperFunctions.replaceWhitespacesFunc(channelNameData);
-			const emojiData = handlerInput.requestEnvelope.request.intent.slots.groupmessageemojiname.value;
-			const emoji = helperFunctions.emojiTranslateFunc(emojiData);
-			const message = messageData + emoji;
-
-			const headers = await helperFunctions.login(accessToken);
-			const roomid = await helperFunctions.getGroupId(channelName, headers);
-			const speechText = await helperFunctions.postGroupMessage(roomid, message, headers);
-			const repromptText = ri('GENERIC_REPROMPT');
-
-
-			return handlerInput.jrb
-				.speak(speechText)
-				.speak(repromptText)
-				.reprompt(repromptText)
-				.withSimpleCard(ri('POST_MESSAGE.CARD_TITLE'), speechText)
-				.getResponse();
-		} catch (error) {
-			console.error(error);
-		}
-	},
-};
-
-const GroupLastMessageIntentHandler = {
-	canHandle(handlerInput) {
-		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
-			handlerInput.requestEnvelope.request.intent.name === 'GroupLastMessageIntent';
-	},
-	async handle(handlerInput) {
-		try {
-			const {
-				accessToken,
-			} = handlerInput.requestEnvelope.context.System.user;
-
-			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.grouplastmessagechannelname.value;
-			const channelName = helperFunctions.replaceWhitespacesFunc(channelNameData);
-
-			const headers = await helperFunctions.login(accessToken);
-			const roomid = await helperFunctions.getGroupId(channelName, headers);
-			const speechText = await helperFunctions.groupLastMessage(channelName, roomid, headers);
-			const repromptText = ri('GENERIC_REPROMPT');
-
-
-			return handlerInput.jrb
-				.speak(speechText)
-				.speak(repromptText)
-				.reprompt(repromptText)
-				.withSimpleCard(ri('GET_LAST_MESSAGE_FROM_CHANNEL.CARD_TITLE'), speechText)
-				.getResponse();
-		} catch (error) {
-			console.error(error);
-		}
-	},
-};
-
-module.exports = {
-	PostGroupEmojiMessageIntentHandler,
-	GroupLastMessageIntentHandler,
-};
