+++ lambda/custom/handlers/General/ChangeStatusIntentHandlers.js
@@ -46,7 +46,14 @@ const ChangeStatusIntentHandler = {
 				.getResponse();
 
 		} catch (error) {
-			console.error(error);
+			const speechText = ri('SOMETHING_WENT_WRONG');
+			const repromptText = ri('GENERIC_REPROMPT');
+
+			return handlerInput.jrb
+				.speak(speechText)
+				.speak(repromptText)
+				.reprompt(repromptText)
+				.getResponse();
 		}
 	},
 };
